+++ src/main/java/com/super_bits/modulosSB/SBCore/UtilGeral/UtilSBCoreGeradorDeID.java
@@ -12,14 +12,14 @@ import org.coletivojava.fw.utilCoreBase.UtilSBCoreGeradorDeIDSimples;
  */
 public abstract class UtilSBCoreGeradorDeID extends UtilSBCoreGeradorDeIDSimples {
 
-    public static int gerarIdUnicoLetrasDaString(String pValor) {
+    public static Long gerarIdUnicoLetrasDaString(String pValor) {
 
         String nomeparaHash = UtilSBCoreStringFiltros.removeCaracteresEspeciaisEspacosETracos(pValor);
         if (UtilSBCoreStringValidador.isNuloOuEmbranco(nomeparaHash)) {
-            return -1;
+            return -1l;
         }
         nomeparaHash = nomeparaHash.toUpperCase();
-        return nomeparaHash.hashCode();
+        return (long) nomeparaHash.hashCode();
     }
 
 }
